% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA-package.R
\docType{package}
\name{PCA-package}
\alias{PCA}
\alias{PCA-package}
\title{PCA: Principle Component Analysis Package}
\description{
The function included in this package lets you explore... Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Erika Carlson \email{erika.carlson@utexas.edu} (\href{https://orcid.org/0000-0002-6638-6378}{ORCID})

Authors:
\itemize{
  \item Kiersten Bell \email{kierstenbell@utexas.edu}
  \item Brody Manquen \email{manquen@utexas.edu}
}

}
\keyword{internal}
