% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAmethod-package.R
\docType{package}
\name{PCAmethod-package}
\alias{PCAmethod}
\alias{PCAmethod-package}
\title{PCAmethod: Principal Component Analysis Package}
\description{
This module is used to build a package to run principal component analysis and visualization using an included dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecarlson5683/PCAmethod}
  \item Report bugs at \url{https://github.com/ecarlson5683/PCAmethod/issues}
}

}
\author{
\strong{Maintainer}: Erika Carlson \email{erika.carlson@utexas.edu} (\href{https://orcid.org/0000-0002-6638-6378}{ORCID})

Authors:
\itemize{
  \item Kiersten Bell \email{kierstenbell@utexas.edu} (\href{https://orcid.org/0000-0002-2073-9342}{ORCID})
  \item Brody Manquen \email{manquen@utexas.edu} (\href{https://orcid.org/0009-0003-7674-4022}{ORCID})
}

}
\keyword{internal}
