% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princomp.R
\name{princomp}
\alias{princomp}
\title{An interactive principle component analysis function}
\usage{
princomp(
  x,
  cor = FALSE,
  scores = TRUE,
  covmat = NULL,
  subset = rep_len(TRUE, nrow(as.matrix(x))),
  ...
)
}
\arguments{
\item{x}{a numeric matrix or data frame with data for PCA}

\item{cor}{a logical value indicating whether the calculation should use the correlation matrix instead of the covariance matrix. Can only use if there are no constant variables. Defaults to FALSE.}

\item{scores}{a logical value indicating whether the score on each principal component should be calculated. Defaults to TRUE.}

\item{covmat}{a covariance matrix or list as returned by `cov.wt`. If TRUE, this is used rather than the covariance matrix of x.}

\item{subset}{an optional vector used to select rows (observations) of the data matrix `x`}

\item{fix_sign}{a logical value indicating whether or not the signs of the loadings and scores should be chosen so that the first element of each loading is non-negative.}
}
\description{
This function allows you to ...
}
\examples{
princomp(x = music_clean)
}
\keyword{pca}
