% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination_utilities.R
\name{unscale}
\alias{unscale}
\title{Some utilities for PCA autoplot}
\usage{
unscale(x, center = NULL, scale = NULL)
}
\description{
These functions assist in PCA autoplots
}
\examples{
match_type(type)
}
\keyword{autoplot}
